tratamento de erros e biulder 201 Create

+++ src/main/resources/db/migration/V1__create-table-parquimetro.sql
@@ -3,9 +3,10 @@ create table parquimetro (
         placa varchar(100) not null,
         dia_hora TIMESTAMP not null,
         entrada TIME not null,
-        saida TIME not null,
+        saida TIME not null ,
         permanecia VARCHAR(255),
-        valor float(3),
+        valor float,
+
 
         primary key(id)
 )
